Rudimentary frame flipbook prototype
A prototype for a frame flipbook for making animated GIFs is now working. The GIF output is not yet connected, but thumbnails can be viewed and the keyframes can be scrolled. First stage animation preview is also working. Several superficial tasks still need to be completed, such as removing highlight from a transport button when the playing state has changed, etc.